requirements.txt
anthropic>=0.40.0
openai>=1.60.0
google-genai>=0.8.0
httpx>=0.27.0
PyYAML>=6.0
numpy>=1.26
scipy>=1.11
jsonschema>=4.21
duckdb>=1.0
python-dateutil>=2.9
tenacity>=8.2
